I think I know the bug.
The bug is when you queue interrupts (keypresses, for example) The
interrupts are enabled first and only then interrupt returns. The
subsequent interrupt however, interrupts the first one before it had
a chance to return, creating this endless loop.
Kind of headfuck, maybe flags and PC should be banked, I don't know.